+++ etna/ensembles/direct_ensemble.py
@@ -1,4 +1,3 @@
-from copy import deepcopy
 from typing import Any
 from typing import Dict
 from typing import List
@@ -100,6 +99,10 @@ class DirectEnsemble(EnsembleMixin, SaveEnsembleMixin, BasePipeline):
     def fit(self, ts: TSDataset, save_ts: bool = True) -> "DirectEnsemble":
         """Fit pipelines in ensemble.
 
+        Method doesn't change the given ``ts``.
+
+        Saved ``ts`` is the link to given ``ts``.
+
         Parameters
         ----------
         ts:
@@ -113,7 +116,7 @@ class DirectEnsemble(EnsembleMixin, SaveEnsembleMixin, BasePipeline):
             Fitted ensemble
         """
         self.pipelines = Parallel(n_jobs=self.n_jobs, **self.joblib_params)(
-            delayed(self._fit_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in self.pipelines
+            delayed(self._fit_pipeline)(pipeline=pipeline, ts=ts) for pipeline in self.pipelines
         )
 
         if save_ts: