+++ etna/ensembles/voting_ensemble.py
@@ -1,4 +1,3 @@
-from copy import deepcopy
 from typing import Any
 from typing import Dict
 from typing import List
@@ -132,7 +131,7 @@ class VotingEnsemble(EnsembleMixin, SaveEnsembleMixin, BasePipeline):
             weights = [1.0 for _ in range(len(self.pipelines))]
         elif self.weights == "auto":
             forecasts = Parallel(n_jobs=self.n_jobs, **self.joblib_params)(
-                delayed(self._backtest_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in self.pipelines
+                delayed(self._backtest_pipeline)(pipeline=pipeline, ts=ts) for pipeline in self.pipelines
             )
 
             x = pd.concat(
@@ -160,6 +159,10 @@ class VotingEnsemble(EnsembleMixin, SaveEnsembleMixin, BasePipeline):
     def fit(self, ts: TSDataset, save_ts: bool = True) -> "VotingEnsemble":
         """Fit pipelines in ensemble.
 
+        Method doesn't change the given ``ts``.
+
+        Saved ``ts`` is the link to given ``ts``.
+
         Parameters
         ----------
         ts:
@@ -173,7 +176,7 @@ class VotingEnsemble(EnsembleMixin, SaveEnsembleMixin, BasePipeline):
             Fitted ensemble
         """
         self.pipelines = Parallel(n_jobs=self.n_jobs, **self.joblib_params)(
-            delayed(self._fit_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in self.pipelines
+            delayed(self._fit_pipeline)(pipeline=pipeline, ts=ts) for pipeline in self.pipelines
         )
         self.processed_weights = self._process_weights(ts=ts)