+++ etna/pipeline/pipeline.py
@@ -1,3 +1,4 @@
+from copy import deepcopy
 from typing import Optional
 from typing import Sequence
 from typing import cast
@@ -54,6 +55,10 @@ class Pipeline(ModelPipelinePredictMixin, ModelPipelineParamsToTuneMixin, SaveMo
 
         Fit and apply given transforms to the data, then fit the model on the transformed data.
 
+        Method doesn't change the given ``ts``.
+
+        Saved ``ts`` is the link to given ``ts``.
+
         Parameters
         ----------
         ts:
@@ -66,9 +71,9 @@ class Pipeline(ModelPipelinePredictMixin, ModelPipelineParamsToTuneMixin, SaveMo
         :
             Fitted Pipeline instance
         """
-        ts.fit_transform(self.transforms)
-        self.model.fit(ts)
-        ts.inverse_transform(self.transforms)
+        cur_ts = deepcopy(ts)
+        cur_ts.fit_transform(self.transforms)
+        self.model.fit(cur_ts)
 
         if save_ts:
             self.ts = ts